payment-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.3
pydantic-settings==2.3.1
PyJWT==2.8.0
pymysql==1.1.1
cryptography>=42.0.0
httpx==0.27.0
